Fix VCN default network detection
VCN-over-wifi causes two network callbacks, one with
TRANSPORT_CELLULAR and a VcnTransportInfo, and another with
TRANSPORT_WIFI with WifiInfo. Currently we match both of these to the
same WifiEntry based on the WifiInfo, but since these have different
Network objects, the entry will flip back and forth between default and
not default (only the Wifi Network is treated as the default).

To fix this, only listen for the network with TRANSPORT_WIFI.

Bug: 329108155
Test: atest WifiPickerTrackerTest
Change-Id: Ied289d08d54d178891d14aba9e8e52317b739dab
(cherry picked from commit aa2e038a43fa11ee16246ec79fbb4b2858227b2d)