Ensure handleOnStart() happens after BaseWifiTracker constructor
BaseWifiTracker constructor takes a Lifecycle argument which may call
onStart() immediately upon registration depending on the implementation.
This can cause handleOnStart() to run on the WorkerThread before the
MainThread finishes the constructor, causing various race conditions.

To fix this, post the lifecycle registration to the main handler so we
run it after the constructor finishes.

Bug: 329642219
Test: atest WifiTrackerLibTests
Change-Id: Iaf9487870ded23fe597fc1b8b897e057e0299ba9