Remove mLock in WifiPickerTracker
mLock is used to synchronize WifiPickerTracker's
getConnectedWifiEntry(), getActiveWifiEntries(), and getWifiEntries()
with the update method updateWifiEntries() on the worker thread.
However, since this update may take a while, the main thread may be
blocked calling the above APIs.

Instead, remove mLock entirely and make the APIs thread safe by not
referencing the return values in the worker thread.

Bug: 326966885
Test: atest WifiPickerTrackerTests
Change-Id: I866731926f58e1454d5a5c62aa77234a5b12d3b5