Notify wifi entries changed when default network changes
SysUI relies on onWifiEntriesChanged() to update the wifi icon to pie or
W+ if W+ is default. However, default network changes don't trigger
onWifiEntriesChanged(), so W+ may be default but not shown in SysUI
until an unrelated onWifiEntriesChanged() triggers. To fix this, make
sure to call onWifiEntriesChanged() even for default network changes.

Bug: 319905522
Test: build
Change-Id: I1ee58955f4af623f99d016a39174e8bf7b7adc72