PANS networks shouldn't make wifi low quality
Cellular per-app default networks may be assigned to be the default
network for a certain set of UIDs determined by the per-app network
selection (PANS) policy, or via 5G slicing. When those UIDs ask if
the WiFi network is "low quality", they are always (incorrectly) told
the WiFi is low quality because the cellular network is their default.

Similarly to VPN, we can infer if the cellular network is set to default
by some policy by checking if it's restricted.

Test: atest WifiPickerTrackerTest#testGetConnectedEntry_wifiValidatedCellDefault_isLowQuality
Bug: 340976866
Change-Id: Ibbd70b3d5a7da20242dd1a915b3d9a7fd693c5e8